Rewrite into a very simple script
The implementation based on eos-ostree-builder is highly customized to
Endless and is unlikely to be used by anyone else. Strip it all down to
a very minimal script that people can test with.

This switches to using multistrap to handle the bootsrapping primarily
because we need to provide a secondary repo that contains the
ostree-boot package. I'm temporarily hosting that at
https://deb-ostree-test.endlessm.com/debian for stretch. A couple
other tweaks are needed to setup for dracut, but this seems to be able
to build a proper system.

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=824649